refactor: rewrite source codes
