+++ src/index.ts
@@ -1,0 +1,2 @@
+export { default as createContext } from "./create-context";
+export { default as useStore } from "./use-store";
